+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 - Improved action and gutter icons to work with high contrast theme
 - Use non-null values in editor state
+- Migrate to Kotlin UI DSL version 2
 - Update platform version to `2021.3.3`
 - `pluginVerifierIdeVersions` – upgrade to `2021.3.3, 2022.1`